break out specific test code into app_boxes_fixtures_test.go
